Move interactive-tx actors to dedicated package
